requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.9.1
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
asyncssh==2.18.0
atpublic==5.0
attrs==23.1.0
billiard==4.2.1
blinker==1.6.2
boto3==1.28.79
botocore==1.31.79
cachetools==5.3.1
celery==5.4.0
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.1.1
cryptography==43.0.3
cycler==0.12.1
dataclasses==0.6
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
Django==4.2.2
dpath==2.2.0
dulwich==0.22.6
dvc==3.56.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==3.0.10
flask-ngrok==0.0.25
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.43.1
frozenlist==1.3.3
fsspec==2024.10.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grandalf==0.8
grpcio==1.56.0
gto==1.7.1
huggingface-hub==0.15.1
hydra-core==1.3.2
idna==3.4
install==1.3.5
iterative-telemetry==0.0.9
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
JPype1==1.4.1
jsonschema==4.17.3
kiwisolver==1.4.5
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@47a69af87928fc24e20f571fe10c3cc9dd9af9a3#subdirectory=kobert_hf
kombu==5.4.2
konlpy==0.6.0
Korpora==0.2.0
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
natsort==8.4.0
networkx==3.1
numpy==1.25.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.1.78
openpyxl==3.1.2
orjson==3.10.11
packaging==23.1
pandas==2.0.2
pathspec==0.12.1
Pillow==10.0.0
platformdirs==3.11.0
prompt_toolkit==3.0.48
protobuf==4.23.3
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyDeprecate==0.3.2
pydot==3.0.2
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
PyMySQL==1.1.0
pyparsing==3.1.1
pyproject-toml==0.0.10
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
pytorch-lightning==1.6.1
pytz==2023.3
pywin32==308
PyYAML==6.0
ratsnlp==1.0.53
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.7.0
scikit-learn==1.3.0
scipy==1.11.1
scmrepo==3.3.8
seaborn==0.13.0
semver==3.0.2
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
sqltrie==0.11.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.13.2
torch==2.0.1
torchdata==0.6.1
torchmetrics==0.11.4
torchtext==0.6.0
torchvision==0.15.2
tqdm==4.65.0
transformers==4.28.1
turtle==0.0.1
typer==0.13.0
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.201
urllib3==1.26.16
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==2.3.6
xlrd==2.0.1
yarl==1.9.2
zc.lockfile==3.0.post1
